% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.aweek.R
\name{as.aweek}
\alias{as.aweek}
\alias{as.aweek.character}
\alias{as.aweek.Date}
\alias{as.aweek.aweek}
\title{Convert characters or dates to aweek objects}
\usage{
as.aweek(x, ...)

\method{as.aweek}{character}(x, start = week_start,
  week_start = get_week_start(), ...)

\method{as.aweek}{Date}(x, ...)

\method{as.aweek}{aweek}(x, week_start = NULL, ...)
}
\arguments{
\item{x}{a correctly formated character or date vector}

\item{...}{(for the \code{Date} method) arguments passed on to \code{\link[=date2week]{date2week()}}}

\item{start}{an integer (or character) vector of days that the weeks
start on for each corresponding week. Defaults to the value of
\code{\link[=get_week_start]{get_week_start()}}. Note that these will not determine the final week.}

\item{week_start}{a number indicating the start of the week based on the ISO
8601 standard from 1 to 7 where 1 = Monday OR an abbreviation of the
weekdate in an English or current locale. \emph{Note: using a non-English locale
may render your code non-portable.} Defaults to the value of
\code{\link[=get_week_start]{get_week_start()}}}
}
\value{
an aweek object
}
\description{
Convert characters or dates to aweek objects
}
\examples{

# aweek objects can only be created from valid weeks:

as.aweek("2018-W10-5", start = 7, week_start = 7) # works!
try(as.aweek("2018-10-5", week_start = 7)) # doesn't work :(

# you can also convert dates
as.aweek(as.Date("2018-03-09"))

# all functions get passed to date2week, so you can use any of its arguments:
as.aweek("2018-W10-5", start = 7, week_start = 7, floor_day = TRUE, factor = TRUE) 
as.aweek(as.Date("2018-03-09"), floor_day = TRUE, factor = TRUE)

# you can convert aweek objects to aweek objects:
x <- get_aweek()
as.aweek(x)
as.aweek(x, week_start = 7)
}
