% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.data.frame.aweek.R
\name{as.data.frame.aweek}
\alias{as.data.frame.aweek}
\title{Convert aweek objects to a data frame}
\usage{
\method{as.data.frame}{aweek}(x, ...)
}
\arguments{
\item{x}{an aweek object}

\item{...}{unused}
}
\value{
a data frame with an aweek column
}
\description{
Convert aweek objects to a data frame
}
\examples{

d  <- as.Date("2019-03-25") + 0:6
w  <- date2week(d, "Sunday")
dw <- data.frame(date = d, week = w)
dw
dw$week
}
\seealso{
\code{\link[=date2week]{date2week()}} \code{\link[=print.aweek]{print.aweek()}}
}
