% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aweek-methods.R
\name{print.aweek}
\alias{print.aweek}
\alias{aweek-class}
\alias{[.aweek}
\alias{[[.aweek}
\alias{[<-.aweek}
\alias{as.list.aweek}
\alias{trunc.aweek}
\alias{rep.aweek}
\alias{c.aweek}
\title{The aweek class}
\usage{
\method{print}{aweek}(x, ...)

\method{[}{aweek}(x, i)

\method{[[}{aweek}(x, i)

\method{[}{aweek}(x, i) <- value

\method{as.list}{aweek}(x, ...)

\method{trunc}{aweek}(x, ...)

\method{rep}{aweek}(x, ...)

\method{c}{aweek}(..., recursive = FALSE, use.names = TRUE)
}
\arguments{
\item{x}{an object of class \code{aweek}}

\item{...}{a series of \code{aweek} objects (unused in \code{print.aweek()})}

\item{i}{index for subsetting an aweek object.}

\item{value}{a value to add or replace in an aweek object}

\item{recursive, use.names}{parameters passed on to \code{\link[=unlist]{unlist()}}}
}
\value{
an object of class \code{aweek}
}
\description{
The methods for combining or modifying aweek objects require
that any aweek obe
}
\details{
The aweek class is a character or factor in the format YYYY-Www(-d) with a
"week_start" attribute containing an integer specifying which day of the ISO
8601 week each week should begin. This documentation shows how to print or
subset the object.
}
\examples{
d <- as.Date("2018-12-20") + 1:40
w <- date2week(d, week_start = "Sunday")
print(w)

# subsetting acts as normal
w[1:10]

# Combining multiple aweek objects will only work if they have the same
# week_start day
c(w[1], w[3], w[5], as.aweek(as.Date("2018-12-01"), week_start = "Sunday"))

# differing week_start days will throw an error
mon <- date2week(as.Date("2018-12-01"), week_start = "Monday")
mon
try(c(w, mon))

# combining Dates will be coerced to aweek objects under the same rules
c(w, Sys.Date())

# truncated aweek objects will be un-truncated
w2 <- date2week(d[1:5], week_start = "Sunday", floor_day = TRUE)
w2
c(w[1:5], w2)
}
\seealso{
\code{\link[=date2week]{date2week()}}, \code{\link[=as.Date.aweek]{as.Date.aweek()}}
}
