% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_week_start.R
\name{change_week_start}
\alias{change_week_start}
\title{Change the week start of an aweek object}
\usage{
change_week_start(x, week_start = get_week_start(), ...)
}
\arguments{
\item{x}{an aweek object}

\item{week_start}{the new desired week_start}

\item{...}{parameters passed to \code{\link[=date2week]{date2week()}}}
}
\description{
This will change the week_start attribute of an aweek object and adjust the
observations accordingly.
}
\examples{
# New Year's 2019 is the third day of the week starting on a Sunday
s <- date2week(as.Date("2019-01-01"), week_start = "Sunday")
s 

# It's the second day of the week starting on a Monday
m <- change_week_start(s, "Monday") 
m

# When you compare the underlying dates, they are exactly the same
identical(as.Date(s), as.Date(m))

# Since this will pass arguments to `date2week()`, you can modify other
# aspects of the aweek object this way, but this is not advised.

change_week_start(s, "Monday", floor_day = TRUE)
}
\seealso{
\code{\link[=as.aweek]{as.aweek()}}, which wraps this function.
}
