% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date2week.R, R/week2date.R
\name{date2week}
\alias{date2week}
\alias{week2date}
\alias{as.Date.rainboweek}
\alias{as.POSIXlt.rainboweek}
\alias{as.character.rainboweek}
\alias{print.rainboweek}
\title{Convert date to a an arbitrary week definition}
\usage{
date2week(x, week_start = 1, floor_day = FALSE, numeric = FALSE, ...)

week2date(x, week_start = 1, floor_day = FALSE)

\method{as.Date}{rainboweek}(x, floor_day = FALSE, ...)

\method{as.POSIXlt}{rainboweek}(x, tz, floor_day = FALSE, ...)

\method{as.character}{rainboweek}(x, ...)

\method{print}{rainboweek}(x, ...)
}
\arguments{
\item{x}{a [Date], [POSIXt], [character], or any data that can be easily
converted to a date with [as.POSIXlt()].}

\item{week_start}{The start of the week from 1 to 7 where 1 = Monday}

\item{floor_day}{when `TRUE`, the days will be set to the start of the week.}

\item{numeric}{if `TRUE`, only the numeric week be returned. If `FALSE`
(default), the date in the format "YYYY-Www-d" will be returned.}

\item{...}{arguments passed to [as.POSIXlt()], unused in all other cases.}

\item{tz}{passed on to [as.POSIXlt()]}
}
\description{
Convert date to a an arbitrary week definition
}
\details{
Weeks differ in their start dates depending on context. The ISO 8601
  standard specifies that Monday starts the week while the US CDC uses Sunday
  as the start of the week. For example, MSF has varying start dates 
  depending on country in order to better coordinate response. 

  While there are package that conver conversion for ISOweeks and epiweeks,
  these do not provide seamless conversion from dates to epiweeks with 
  non-standard start dates. This package provides a lightweight utility to
  be able to convert each day.
}
\examples{

# The same set of days will occur in different weeks depending on the start
# date. Here we can define a week before and after today

print(dat <- Sys.Date() + -6:7)

# By default, the weeks are defined as ISO weeks, which start on Monday
print(iso_dat <- date2week(dat))

# If you want lubridate-style numeric-only weeks, you need look no further
# than the "numeric" argument
date2week(dat, 1, numeric = TRUE)

# The rainboweek class can be converted back to a date with `as.Date()`
as.Date(iso_dat)

# If you want to show only the first day of the week, you can use the 
# `floor_day` argument
as.Date(iso_dat, floor_day = TRUE)

# ISO week definition: Monday -- 1
date2week(dat, 1)

# Tuesday -- 2
date2week(dat, 2)

# Wednesday -- 3
date2week(dat, 3)

# Thursday -- 4
date2week(dat, 4)

# Friday -- 5
date2week(dat, 5)

# Saturday -- 6
date2week(dat, 6)

# Epiweek definition: Sunday -- 7 
date2week(dat, 7)
}
\author{
Zhian N. Kamvar
}
